Email slicer
Email Slicer is just a simple tool that will take multiple email address as an input and slice it to produce the username and the domain associated with it. The email must be divided into two strings by using ‘@’ as the separator.
So, user provides n number of email addresses and you have to design a logic to slice the username and the domain out of those email addresses. The domain part must print in capitals.